% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaq.R
\name{char2ic}
\alias{char2ic}
\title{Convert character input to labeled item codes}
\usage{
char2ic(i)
}
\arguments{
\item{i}{character vector of item labels.}
}
\description{
Convert character input to labeled item codes
}
\keyword{internal}
