% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaq.R
\name{croaq}
\alias{croaq}
\alias{summary.croaq}
\alias{plot.croaq}
\title{Fetch air quality data}
\usage{
croaq(station, item, from, to, type = 0, labels = NULL, fill = TRUE)

\method{summary}{croaq}(object, ...)

\method{plot}{croaq}(x, ...)
}
\arguments{
\item{station}{integer, station code.}

\item{item}{integer or character, pollutant codes or labels.}

\item{from, to}{character, start and end date.}

\item{type}{integer, value datatype.}

\item{labels}{character, custom item labels.}

\item{fill}{logical, if \code{TRUE} (default) fill data within time range with missing values where no data is available.}

\item{object}{\code{croaq} object.}

\item{...}{further parameters.}

\item{x}{\code{croaq} object.}
}
\value{
Object of class \code{croaq} inheriting from a data frame class with dedicated summary and plot methods.

Summary method invisibly returns total and missing values for each item.

Plot method shows time on x-axis and values on y-axis. Time points with missing values are shown as vertical red lines.
}
\description{
Fetch air quality data from state and local networks of measurement stations in Croatia.
}
\details{
Fetches values of one or more \code{item}s (pollutants, etc.) measured at a single \code{station} within time period specified by \code{from} and \code{to}.

For a list of stations and their codes, run \code{listStations()}.

Items can be specified using integer codes or (in some cases) short labels, run \code{listPollutants()} for a list. Custom labels can be set by \code{labels} argument which should be a character vector of the same length as \code{item} and have unique values.

There are various data\code{types} available (raw, validated, hourly, daily, etc.). Run \code{listDatatypes()} for a list.

Dates should be specified as character values in `Y-m-d` format. If \code{fill=TRUE} (default), time points with missing values will be added to data where no data is available. If \code{fill = FALSE}, missing values are not accounted for in summary and plot methods.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Summary of \code{croaq} object

\item \code{plot}: Plot \code{croaq} object.
}}

\note{
Availability of data depends on selected station, time period and datatype. A note is printed if requested data for some \code{item}s is not available and if no data at all is available \code{NULL} is returned.
}
\references{
Data exchange service at \url{http://iszz.azo.hr/iskzl/exc.htm}.
}
\seealso{
\code{\link{wide}}
}
