% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaq.R
\name{fill_time_range}
\alias{fill_time_range}
\title{Fill time range with missing values}
\usage{
fill_time_range(data, time.range, type)
}
\arguments{
\item{data}{data frame.}

\item{time.range}{vector of length 2 in POSIXct format.}

\item{type}{integer, datatype.}
}
\description{
Fill time range with missing values
}
\keyword{internal}
