% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaq.R
\name{listPollutants}
\alias{listPollutants}
\title{List items (pollutants)}
\usage{
listPollutants()
}
\value{
A data frame with items (pollutants) listing.
}
\description{
List items (pollutants)
}
\references{
\url{http://iszz.azo.hr/iskzl/doc/servis_uputa.docx}
}
