% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaq.R
\name{listStations}
\alias{listStations}
\title{List stations}
\usage{
listStations()
}
\value{
A data frame with stations listing.
}
\description{
List stations
}
\references{
\url{http://iszz.azo.hr/iskzl/doc/servis_uputa.docx}
}
