% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaq.R
\name{num2ic}
\alias{num2ic}
\title{Convert numeric/integer input to labeled item codes}
\usage{
num2ic(i)
}
\arguments{
\item{i}{numeric/integer vector of item codes.}
}
\description{
Convert numeric/integer input to labeled item codes
}
\keyword{internal}
