% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaq.R
\name{set_items}
\alias{set_items}
\title{Check and set item specification.}
\usage{
set_items(item, labels)
}
\arguments{
\item{item}{integer or character.}

\item{labels}{character.}
}
\description{
Check and set item specification.
}
\keyword{internal}
