% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaq.R
\name{set_time_range}
\alias{set_time_range}
\title{Check and set time range}
\usage{
set_time_range(from, to)
}
\arguments{
\item{from, to}{character.}
}
\description{
Check and set time range
}
\keyword{internal}
