% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croaq.R
\name{wide}
\alias{wide}
\title{Reshape data into wide format}
\usage{
wide(data, idvar = c("station", "time", "type"), timevar = "label")
}
\arguments{
\item{data}{object of class \code{croaqData}.}

\item{idvar}{see \code{\link[stats]{reshape}}.}

\item{timevar}{see \code{\link[stats]{reshape}}.}
}
\description{
Sometimes it is handy to work with data in wide format.
}
